refactor: extract cli args, config, store modules and authoring verbs to commands/
