+++ cli/src/commands/mod.rs
@@ -1,0 +1,5 @@
+//! One module per CLI subcommand wrapper, mirroring `living_docs_core::commands` (issue 0028).
+
+pub(crate) mod brief;
+pub(crate) mod new;
+pub(crate) mod next;
